mongo: init feedback db
I want to reply the feedbacks without caring the user ids. That's
why message's and related user's id will be stored in the database.
Whenever I "reply" a message it will find the related users and send
my message to them.

Telegram actually provides user_id when you forward a message, but
this is only for the public accounts. Most of the users, probably
by default, don't turn on this feature and it prevents us to see
user_id. Even though bot is the one who forwarded the message.
Otherwise we wouldn't need a database to store these ids.