handler: handle /add and /remove with ConversationHandler
Resolves #44

In this commit, I also had to introduce grouping for the handlers,
because jumping from one conversation to another conversation or
command is very much likely among the users. Fixes #42.

locale: add new translations for the /add and /remove commands.

While rewriting the add and remove handlers, it occured to me
making `subscriptions` could be nice. MongoDB doesn't support sets
atm, but even though making it might be more efficient. I have to
think this later.