fix.scraper: don't encode "&" in while fixing invalid url
Some announcement urls are weird. They make fancy queries in the
website (honestly sometimes the whole website runs with http
parameters lol). I realized when we encode "&" the fixed url does
not work in this type of situtations.

I tested a few other weird urls, like the ones who have html tags
in it. And it was working alright. I hope I don't break something
with this fix, but let's see.